requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langgraph>=0.0.20

# Vector stores
faiss-cpu>=1.7.4

# Document processing
beautifulsoup4>=4.12.0
pypdf>=3.17.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Web search
duckduckgo-search>=4.0.0

# Additional dependencies for v1.1.0
networkx>=3.2
matplotlib>=3.8.0
plotly>=5.18.0
sqlalchemy>=2.0.25
pandas>=2.2.0
spacy>=3.7.0
ragas>=0.1.7
datasets>=2.16.0

# New dependencies for v1.2.0 (Multimodal RAG)
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.27.0
